fix(sync): ReadDuration changed to int64. (#24)
This will fix the uploading issue some user have, due to int32.

+++ internal/domain/mdata.go
@@ -79,7 +79,7 @@ type Tracking struct {
 type History struct {
 	URL          string `json:"url"`
 	LastRead     int64  `json:"lastRead"`
-	ReadDuration int    `json:"readDuration"`
+	ReadDuration int64  `json:"readDuration"`
 }
 
 type Sources struct {
